+++ packages/grid/src/BaseCell.css
@@ -52,11 +52,11 @@
 }
 
 .uitkGridBaseCell-focused {
-  border-style: dotted;
-  border-width: 2px;
+  border-style: var(--grid-cursor-border-style);
+  border-width: var(--grid-cursor-border-width);
   border-color: var(--grid-cursor-borderColor);
-  line-height: calc(var(--grid-row-height) - 4px);
-  --grid-cell-padding: 0 calc(var(--grid-padding) - 2px);
+  line-height: calc(var(--grid-row-height) - 2 * var(--grid-cursor-border-width));
+  --grid-cell-padding: 0 calc(var(--grid-padding) - var(--grid-cursor-border-width));
 }
 
 .uitkGridBaseCell-editable:hover {
@@ -70,7 +70,7 @@
   width: 0;
   height: 0;
   border-top: 0 solid transparent;
-  border-left: 8px solid var(--grid-cursor-borderColor);
+  border-left: var(--grid-editableCell-cornerTag-size) solid var(--grid-cursor-borderColor);
   border-bottom: 8px solid transparent;
   z-index: 10;
 }
