+++ packages/grid/src/GroupHeaderCellValue.css
@@ -1,0 +1,2 @@
+.uitkGridGroupHeaderCellValue {
+}
