+++ packages/grid/src/index.ts
@@ -1,6 +1,8 @@
 export * from "./BaseCell";
 export * from "./ColumnGroup";
+export * from "./CellEditor";
 export * from "./GroupHeaderCell";
+export * from "./GroupHeaderCellValue";
 export * from "./HeaderCell";
 export * from "./HeaderCellValue";
 export * from "./NumberRange";
