+++ packages/grid/stories/examples/ControlledRowSelectionExample.tsx
@@ -1,0 +1,56 @@
+import {
+  Grid,
+  GridColumn,
+  GridProps,
+  RowSelectionCheckboxColumn,
+} from "../../src";
+import { DummyRow, dummyRowKeyGetter, rowData } from "./GridVariantExample";
+import { useState } from "react";
+import { FlexLayout } from "@jpmorganchase/uitk-core";
+import "./Examples.css";
+
+export const ControlledRowSelectionExample = (
+  props: Partial<GridProps<DummyRow>>
+) => {
+  const [selection, setSelection] = useState<number[]>([]);
+
+  const onRowSelected = (rowIndices: number[]) => {
+    setSelection(rowIndices);
+  };
+
+  return (
+    <FlexLayout direction="row">
+      <Grid
+        rowData={rowData}
+        rowKeyGetter={dummyRowKeyGetter}
+        className="grid-small"
+        cellSelectionMode="none"
+        columnSeparators={true}
+        zebra={true}
+        rowSelectionMode="multi"
+        selectedRowIdxs={selection}
+        onRowSelected={onRowSelected}
+        {...props}
+      >
+        <RowSelectionCheckboxColumn id="checkbox" />
+        <GridColumn name="A" id="a" defaultWidth={50} getValue={(r) => r.a} />
+      </Grid>
+      <Grid
+        rowData={rowData}
+        rowKeyGetter={dummyRowKeyGetter}
+        className="grid-small"
+        cellSelectionMode="none"
+        columnSeparators={true}
+        zebra={true}
+        rowSelectionMode="multi"
+        selectedRowIdxs={selection}
+        onRowSelected={onRowSelected}
+        {...props}
+      >
+        <RowSelectionCheckboxColumn id="checkbox" />
+        <GridColumn name="A" id="a" defaultWidth={50} getValue={(r) => r.a} />
+        <GridColumn name="B" id="b" defaultWidth={50} getValue={(r) => r.b} />
+      </Grid>
+    </FlexLayout>
+  );
+};
