+++ packages/grid/stories/examples/RowSelectionModesExample.tsx
@@ -1,0 +1,81 @@
+import { Story } from "@storybook/react";
+import {
+  Grid,
+  GridColumn,
+  GridProps,
+  GridRowSelectionMode,
+  NumericColumn,
+  RowSelectionCheckboxColumn,
+  RowSelectionRadioColumn,
+} from "../../src";
+import { DummyRow, dummyRowKeyGetter, rowData } from "./GridVariantExample";
+import { useState } from "react";
+import {
+  ToggleButton,
+  ToggleButtonGroup,
+  ToggleButtonGroupChangeEventHandler,
+} from "@jpmorganchase/uitk-lab";
+import { FlexItem, FlexLayout } from "@jpmorganchase/uitk-core";
+import "./Examples.css";
+
+export const RowSelectionModesExample = (
+  props: Partial<GridProps<DummyRow>>
+) => {
+  const rowSelectionModes: GridRowSelectionMode[] = ["multi", "single", "none"];
+  const [index, setIndex] = useState<number>(0);
+
+  const onChange: ToggleButtonGroupChangeEventHandler = (
+    event,
+    index,
+    toggled
+  ) => {
+    setIndex(index);
+  };
+
+  const rowSelectionMode = rowSelectionModes[index];
+
+  return (
+    <FlexLayout direction="column">
+      <FlexItem>
+        <ToggleButtonGroup onChange={onChange} selectedIndex={index}>
+          <ToggleButton ariaLabel="multi" tooltipText="Multi">
+            Multi
+          </ToggleButton>
+          <ToggleButton ariaLabel="single" tooltipText="Single">
+            Single
+          </ToggleButton>
+          <ToggleButton ariaLabel="none" tooltipText="None">
+            None
+          </ToggleButton>
+        </ToggleButtonGroup>
+      </FlexItem>
+      <Grid
+        rowData={rowData}
+        rowKeyGetter={dummyRowKeyGetter}
+        className="grid"
+        cellSelectionMode="none"
+        columnSeparators={true}
+        zebra={true}
+        {...props}
+        rowSelectionMode={rowSelectionMode}
+      >
+        {rowSelectionMode === "multi" && (
+          <RowSelectionCheckboxColumn id="checkbox" />
+        )}
+        {rowSelectionMode === "single" && (
+          <RowSelectionRadioColumn id="radio" />
+        )}
+        <GridColumn name="A" id="a" defaultWidth={50} getValue={(r) => r.a} />
+        <NumericColumn
+          name="B"
+          id="b"
+          defaultWidth={100}
+          precision={2}
+          getValue={(r: DummyRow) => r.b}
+          align={"right"}
+        />
+        <GridColumn name="C" id="c" defaultWidth={50} getValue={(r) => r.c} />
+      </Grid>
+    </FlexLayout>
+  );
+};
